fix: close ad30 follow-up findings
