+++ docs/plans/phase-AD/readiness.md
@@ -18,7 +18,7 @@ Companion closure ledger:
 
 Current readiness verdict:
 
-- `release verdict: pending AD.30 final close/not-close decision`
+- `release verdict: pending CI confirmation for the terminal AD.30 fix round`
 
 Current evidence surfaces:
 
@@ -32,9 +32,9 @@ Current evidence surfaces:
 | `AD.25` | `complete` | `feature/pAD-s25-post-send-hook-emitter-live-wiring` | `../atm-core-worktrees/feature/pAD-s25-post-send-hook-emitter-live-wiring` | override rows expose explicit override, disable, and clear/reset semantics with one stable empty-body error contract |
 | `AD.26` | `complete` | `feature/pAD-s26-rule001-observability-seam-closure` | `../atm-core-worktrees/feature/pAD-s26-rule001-observability-seam-closure` | `PostSendHookEmitter` and `GraftPostSendPort` are live seams, mixed-success accounting is real, and the `RULE-001` library-internal adapter exception is closed through `ADR-020` plus lint enforcement so only the sanctioned daemon adapter module keeps the direct observability imports |
 | `AD.27` | `complete` | `feature/pAD-s27-upstream-built-in-template-resolution` | `../atm-core-worktrees/feature/pAD-s27-upstream-built-in-template-resolution` | `ADR-019-EXC-AD26-001` is closed and built-in override lookup is upstream of `PostSendHookEmitter` |
-| `AD.28` | `planned` | `feature/pAD-s28-atm-graft-timing-independent` | `../atm-core-worktrees/feature/pAD-s28-atm-graft-timing-independent` | the graft host-nudge readiness race is closed through deterministic readiness, not timeout luck |
-| `AD.29` | `planned` | `feature/pAD-s29-phase-ad-post-send-smoke-matrix` | `../atm-core-worktrees/feature/pAD-s29-phase-ad-post-send-smoke-matrix` | one authoritative smoke lane proves the repaired post-send matrix |
-| `AD.30` | `in_progress` | `feature/pAD-s30-windows-daemon-integration-depth` | `../atm-core-worktrees/feature/pAD-s30-windows-daemon-integration-depth` | Windows local-IPC depth coverage is restored, the direct-fix ledger is closed, and this readiness file records the final verdict |
+| `AD.28` | `complete` | `feature/pAD-s28-atm-graft-timing-independent` | `../atm-core-worktrees/feature/pAD-s28-atm-graft-timing-independent` | the graft host-nudge readiness race is closed through deterministic readiness, not timeout luck |
+| `AD.29` | `complete` | `feature/pAD-s29-phase-ad-post-send-smoke-matrix` | `../atm-core-worktrees/feature/pAD-s29-phase-ad-post-send-smoke-matrix` | one authoritative smoke lane proves the repaired post-send matrix |
+| `AD.30` | `complete` | `feature/pAD-s30-windows-daemon-integration-depth` | `../atm-core-worktrees/feature/pAD-s30-windows-daemon-integration-depth` | Windows local-IPC depth coverage is restored, the direct-fix ledger is closed, and this readiness file records the final verdict |
 
 ## Direct-Fix Carry-Forward Ledger
